Refactored Pickup item system
Pickup item system is more flexible, since I used abstraction and interfaces in order to make this system reusable.

Now, I can simply add new SO to create new items of the same type or change settings in SO of that item only in editor. Moreover I can easily add new item types. This can be done by adding a new class with mix of interfaces and one required abstract class for inheritance.