Implementing flexible event system in project
Started to replacing GameManager singleton calls with ScriptableObjects as an intermediate point of communication.

So far I added VoidChannelSO, which contains Action event and RaiseEvent method for calling events. Each of systems that want to do sth in specific moment can subscribe to this event on SO.

This channel system based on events can be widely used also in other systems like UI or audio.